LtcPay/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
httpx==0.28.1
redis==5.2.1
jinja2==3.1.5
python-multipart==0.0.20
alembic==1.14.1
bcrypt==4.2.1
gunicorn==23.0.0
email-validator==2.2.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.19.2
